+++ apis/v1alpha1/cluster_types.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2023 The Cockroach Authors
+Copyright 2025 The Cockroach Authors
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -145,6 +145,11 @@ type CrdbClusterSpec struct {
 	// Default: false
 	// +optional
 	AutomountServiceAccountToken bool `json:"automountServiceAccountToken,omitempty"`
+	// (Optional) The grace period in seconds prior to the container being forcibly terminated
+	// when marked for deletion or restarted.
+	// Default : 300
+	// +optional
+	TerminationGracePeriodSecs int64 `json:"terminationGracePeriodSecs,omitempty"`
 }
 
 // +k8s:openapi-gen=true