+++ e2e/upgrades/upgrades_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2023 The Cockroach Authors
+Copyright 2025 The Cockroach Authors
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -28,6 +28,21 @@ import (
 	"github.com/go-logr/zapr"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap/zaptest"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
+)
+
+var (
+	resRequirements = corev1.ResourceRequirements{
+		Limits: corev1.ResourceList{
+			corev1.ResourceCPU:    resource.MustParse(e2e.DefaultCPULimit),
+			corev1.ResourceMemory: resource.MustParse(e2e.DefaultMemoryLimit),
+		},
+		Requests: corev1.ResourceList{
+			corev1.ResourceCPU:    resource.MustParse(e2e.DefaultCPURequest),
+			corev1.ResourceMemory: resource.MustParse(e2e.DefaultMemoryRequest),
+		},
+	}
 )
 
 // TestUpgradesMinorVersion tests a minor version bump
@@ -52,7 +67,7 @@ func TestUpgradesMinorVersion(t *testing.T) {
 
 	builder := testutil.NewBuilder("crdb").WithNodeCount(3).WithTLS().
 		WithImage(e2e.MinorVersion1).
-		WithPVDataStore("1Gi")
+		WithPVDataStore("1Gi").WithResources(resRequirements)
 
 	steps := testutil.Steps{
 		{
@@ -103,7 +118,7 @@ func TestUpgradesMajorVersion20to21(t *testing.T) {
 
 	builder := testutil.NewBuilder("crdb").WithNodeCount(3).WithTLS().
 		WithImage(e2e.MinorVersion2).
-		WithPVDataStore("1Gi")
+		WithPVDataStore("1Gi").WithResources(resRequirements)
 
 	steps := testutil.Steps{
 		{
@@ -152,7 +167,7 @@ func TestUpgradesMajorVersion20_1To20_2(t *testing.T) {
 
 	builder := testutil.NewBuilder("crdb").WithNodeCount(3).WithTLS().
 		WithImage("cockroachdb/cockroach:v20.1.16").
-		WithPVDataStore("1Gi")
+		WithPVDataStore("1Gi").WithResources(resRequirements)
 
 	steps := testutil.Steps{
 		{
@@ -173,6 +188,53 @@ func TestUpgradesMajorVersion20_1To20_2(t *testing.T) {
 				require.NoError(t, sb.Patch(updated, client.MergeFrom(current)))
 				// we wait 10 min because we will be waiting 3 min for each pod because
 				// v20.1.16 does not have curl installed
+				testutil.RequireClusterToBeReadyEventuallyTimeout(t, sb, builder, e2e.CreateClusterTimeout)
+				testutil.RequireDbContainersToUseImage(t, sb, updated)
+				t.Log("Done with major upgrade")
+			},
+		},
+	}
+
+	steps.Run(t)
+}
+
+func TestUpgradesMajorVersionSkippingInnovativeRelease24_3To25_1(t *testing.T) {
+
+	if testing.Short() {
+		t.Skip("skipping test in short mode.")
+	}
+
+	testLog := zapr.NewLogger(zaptest.NewLogger(t))
+
+	e := testenv.CreateActiveEnvForTest()
+	env := e.Start()
+	defer e.Stop()
+
+	sb := testenv.NewDiffingSandbox(t, env)
+	sb.StartManager(t, controller.InitClusterReconcilerWithLogger(testLog))
+
+	builder := testutil.NewBuilder("crdb").WithNodeCount(3).WithTLS().
+		WithImage("cockroachdb/cockroach:v24.3.4").
+		WithPVDataStore("1Gi").WithResources(resRequirements)
+
+	steps := testutil.Steps{
+		{
+			Name: "creates a 3-node secure cluster",
+			Test: func(t *testing.T) {
+				require.NoError(t, sb.Create(builder.Cr()))
+				testutil.RequireClusterToBeReadyEventuallyTimeout(t, sb, builder, e2e.CreateClusterTimeout)
+			},
+		},
+		{
+			Name: "upgrades the cluster to the next major version skipping innovative release",
+			Test: func(t *testing.T) {
+				current := builder.Cr()
+				require.NoError(t, sb.Get(current))
+
+				updated := current.DeepCopy()
+				updated.Spec.Image.Name = "cockroachdb/cockroach:v25.1.0"
+				require.NoError(t, sb.Patch(updated, client.MergeFrom(current)))
+				// we wait 10 min because we will be waiting 3 min for each pod.
 				testutil.RequireClusterToBeReadyEventuallyTimeout(t, sb, builder, e2e.CreateClusterTimeout)
 				testutil.RequireDbContainersToUseImage(t, sb, updated)
 				t.Log("Done with major upgrade")
@@ -206,10 +268,12 @@ func TestUpgradesMinorVersionThenRollback(t *testing.T) {
 
 	builder := testutil.NewBuilder("crdb").
 		WithAutomountServiceAccountToken(true).
+		WithTerminationGracePeriodSeconds(5).
 		WithNodeCount(3).
 		WithTLS().
 		WithImage(e2e.MinorVersion1).
-		WithPVDataStore("1Gi")
+		WithPVDataStore("1Gi").
+		WithResources(resRequirements)
 
 	steps := testutil.Steps{
 		{
@@ -274,7 +338,8 @@ func TestUpgradeWithInvalidVersion(t *testing.T) {
 
 	builder := testutil.NewBuilder("crdb").WithNodeCount(3).WithTLS().
 		WithImage(e2e.MinorVersion1).
-		WithPVDataStore("1Gi")
+		WithPVDataStore("1Gi").
+		WithResources(resRequirements)
 
 	steps := testutil.Steps{
 		{
@@ -324,7 +389,8 @@ func TestUpgradeWithInvalidImage(t *testing.T) {
 
 	builder := testutil.NewBuilder("crdb").WithNodeCount(3).WithTLS().
 		WithImage(e2e.MinorVersion1).
-		WithPVDataStore("1Gi")
+		WithPVDataStore("1Gi").
+		WithResources(resRequirements)
 
 	steps := testutil.Steps{
 		{
@@ -374,7 +440,7 @@ func TestUpgradeWithMajorVersionExcludingMajorFeature(t *testing.T) {
 
 	builder := testutil.NewBuilder("crdb").WithNodeCount(3).WithTLS().
 		WithImage(e2e.SkipFeatureVersion).
-		WithPVDataStore("1Gi")
+		WithPVDataStore("1Gi").WithResources(resRequirements)
 
 	steps := testutil.Steps{
 		{