+++ hack/dev.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-# Copyright 2023 The Cockroach Authors
+# Copyright 2025 The Cockroach Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -21,7 +21,8 @@ APP_VERSION="${APP_VERSION:-v$(cat version.txt)}"
 CLUSTER_NAME="dev"
 NODE_IMAGE="rancher/k3s:v1.23.3-k3s1"
 REGISTRY_NAME="registry.localhost"
-REGISTRY_PORT=5000
+REGISTRY_PORT=5002
+DEV_REGISTRY="${REGISTRY_NAME}:${REGISTRY_PORT}"
 
 main() {
   case "${1:-}" in
@@ -45,13 +46,18 @@ install_operator() {
   # Can't seem to figure out how to leverage the stamp variables here. So for
   # now I've added a defined make variable which can be used for substitution
   # in //config/default/BUILD.bazel.
+  # ${REGISTRY_NAME} must be mapped to localhost or 127.0.0.1 to push the image to the k3d registry.
+  bazel run //hack/crdbversions:crdbversions -- -operator-image ${DEV_REGISTRY}/cockroach-operator -operator-version ${APP_VERSION} -crdb-versions $(PWD)/crdb-versions.yaml -repo-root $(PWD)
   K8S_CLUSTER="k3d-${CLUSTER_NAME}" \
-    DEV_REGISTRY="${REGISTRY_NAME}:${REGISTRY_PORT}" \
+    DOCKER_REGISTRY="${DEV_REGISTRY}" \
+    DOCKER_IMAGE_REPOSITORY="cockroach-operator" \
+    APP_VERSION="${APP_VERSION}" \
     bazel run \
     --stamp \
-    --platforms=@io_bazel_rules_go//go/toolchain:linux_amd64 \
-    --define APP_VERSION=${APP_VERSION} \
-    //config/default:install.apply
+    //:push_operator_image
+    bazel run \
+        --stamp \
+    //config/default:apply_k8s_app
 }
 
 wait_for_ready() {