+++ pkg/resource/cluster.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2023 The Cockroach Authors
+Copyright 2025 The Cockroach Authors
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -328,6 +328,13 @@ func (cluster Cluster) SecureMode() string {
 	}
 
 	return "--insecure"
+}
+
+func (cluster Cluster) GetTerminationGracePeriod() int64 {
+	if cluster.Spec().TerminationGracePeriodSecs == 0 {
+		return terminationGracePeriodSecs
+	}
+	return cluster.Spec().TerminationGracePeriodSecs
 }
 
 func (cluster Cluster) LoggingConfiguration(fetcher Fetcher) (string, error) {