+++ pkg/update/internal.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2023 The Cockroach Authors
+Copyright 2025 The Cockroach Authors
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -23,6 +23,16 @@ const (
 	RootSQLUser = "root"
 	NodeUser    = "node"
 	AdminRole   = "admin"
+)
+
+// ReleaseType represents the type of a release
+type ReleaseType int
+
+const (
+	// Regular releases are mandatory upgrades that cannot be skipped.
+	Regular ReleaseType = iota
+	// Innovative releases are optional upgrades that can be skipped.
+	Innovative
 )
 
 // internalUsers is a set of SQL users created as part of the managed service, not to be used